Refactor astrometry.net backend into anet submodule
