+++ skylib/astrometry/anet/ngc2000.dat
